backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
telethon
celery
redis
python-dotenv
